Generate completions at build time
